Added specific port functionality along with some http parsing functions
This is the last commit to document what I did for the java part.
It was great but short lived indeed. I did not like any serverside API
for java. So I will move on with the lovely node.js, I did try to avoid
it before just cause I wanted to try java.

This commit has the latest state of what could potentially be a
loadbalncer application that can listen on multple ports and get
deployed on serveral machines. I know there are probably way better APIs
for doing this kind of stuff but it was worth the learning journy, or
not, only time can tell. But for now, time will tell me and java apart.
# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
#
# On branch master
# Changes to be committed:
#	modified:   makefile
#	modified:   src/Hangman.class
#	modified:   src/Hangman.java
#	modified:   src/Network/ClientListner.class
#	modified:   src/Network/ClientListner.java
#	modified:   src/Network/Listner.class
#	modified:   src/Network/Listner.java
#	modified:   src/Network/Listners.class
#	modified:   src/Network/Listners.java
#
# Untracked files:
#	expressApp/
#