Managed to remove some nouns from the db
So with some magic I managed to get rid of around 10000 nouns from the
db which should make playing the game better. I also added a fail-safe
mesure when word doesn't get initialized from db in time. What I
baiscally did is set it to the same word it was before.
# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
#
# On branch master
# Your branch is up-to-date with 'origin/master'.
#
# Changes to be committed:
#	modified:   create.sql
#	modified:   load.txt
#	modified:   routes/index.js
#	modified:   words
#