count loses when reseting+fail safe db query+congrats/lose sentences
So now the logic will count reseting as losing a point no matter what,
that way poeple won't cheat.
Also when losing/reseting it will display the word till a guess is made
in the new automatically generated game.
It will congrts the user when winning.
Lastly, I added an array of 20 words as a fail safe for when the
database takes too long to get back with a query, which when tested on
my machine happened more often than I thought. I happens way more when
the code runs in production given that I do not console log anything in
production except which port the app is listing on, boosting the
performance of the app a lot, What happens is that the async call
doesn't return in time for it to be read. This is probably also due to other
optimizaitons done in production.
# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
#
# On branch master
# Your branch is ahead of 'origin/master' by 1 commit.
#   (use "git push" to publish your local commits)
#
# Changes to be committed:
#	modified:   routes/index.js
#	modified:   views/index.hbs
#